# eslint warn floating promises