--- a/packages/governance/package.json
+++ b/packages/governance/package.json
@@ -63,11 +63,6 @@
     ],
     "timeout": "10m"
   },
-  "eslintConfig": {
-    "extends": [
-      "@agoric"
-    ]
-  },
   "publishConfig": {
     "access": "public"
   }