--- a/packages/wallet/ui/package.json
+++ b/packages/wallet/ui/package.json
@@ -81,7 +81,8 @@
       "react/prop-types": "off",
       "react/react-in-jsx-scope": "off",
       "import/no-extraneous-dependencies": "off",
-      "react/display-name": "off"
+      "react/display-name": "off",
+      "@typescript-eslint/no-floating-promises": "off"
     },
     "env": {
       "browser": true,