closes: #5407

## Description

Enable [no-floating-promises](https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/eslint-plugin/docs/rules/no-floating-promises.md). This defers solving all the floating promises by using `warn` level instead of `error`. Eventually we'll want to `error` but I propose we get there incrementally, distributed among teams.

This [rule from @typescript-eslint](https://typescript-eslint.io/rules/) is the first we've enabled to require type information (by setting `parserOptions.project`) . That increases the time cost of linting. Measured with `time yarn lerna run --no-bail lint:eslint` on my machine:
- before: 84.86s user 5.35s system 454% cpu 19.827 total (1.4min)
- after: 824.27s user 135.32s system 556% cpu 2:52.39 total (14min)

In CI the lint job went from [5 min before](https://github.com/Agoric/agoric-sdk/runs/6693643820?check_suite_focus=true) to [21 min after](https://github.com/Agoric/agoric-sdk/runs/6694898124?check_suite_focus=true). To prevent developers having to wait longer for PRs to complete, this splits the lint job into two `lint-primary` and `lint-rest`. 

The comments in that job config note that performance could be greatly improved if the linter were able to [reuse type information across projects](https://github.com/typescript-eslint/typescript-eslint/issues/2094). Alternately we could [unify the packages into one TypeScript project](https://github.com/Agoric/agoric-sdk/pull/4404).

I also tried cutting the lint job time by shortening the setup step, replacing `restore-node` with `setup-node` and yarn install from cache. This cut [5min setup time](https://github.com/Agoric/agoric-sdk/runs/6710173722?check_suite_focus=true) down to [2min](https://github.com/Agoric/agoric-sdk/runs/6710773587?check_suite_focus=true).  (Saving 6min compute over the two jobs.) But it failed because there are build artifacts the linting depends on, such as:
- eslint `import/no-unresolved` fails when `@agoric/ui-components` isn't built
- typescript cannot find module '../bundles/bundle-contractFacet.js' in ./zoe/tools/fakeVatAdmin.js (maybe this one should use bundleCache instead?)

### Security Considerations

--

### Documentation Considerations

Notify developers that they'll begin seeing new warnings and that the Github PR annotations may show many more in the code review web UI.

### Testing Considerations

Verified that lint jobs are not the slowest zebra: iIn last run lint-primary (10m) and lint-rest (12m) were both less than test-cosmic-swingset (13m) and test-swingset3 (13min).