@turadg 2022-06-04
> Running yarn lint locally took 33 minutes and failed with an OOM in the solo package.

Recording numbers from out chat: (in packages/zoe)
mhofman's VM: 99s before, 1221s after (~12x slowdown)
turadg's M1: 5s before, 42s after (~8x slowdown)

> Let's find a way to reduce the impact of this change.

👍  [e1625fe](https://github.com/Agoric/agoric-sdk/pull/5495/commits/e1625fec12081dbe9cc9708e50c975f9784ba26d) makes it opt in with an `AGORICSDK_ESLINT_ENABLE_TYPE_RULES` flag. CI will use this and those who want to use it locally can enable it too.