@turadg 2022-06-06
> Should we also add a new script to the root package.json that runs lint:packages

All the `lint:*` packages get run by `yarn lint` so that would duplicate. We could have a name in the root doesn't start with `lint:` but in package.json we'd lose the ability to document why the env vars are necessary. We could have a separate script, but I find the existing solution sufficient. It works for CI. On my machine I'll set the env var. On yours you know what to do. On others they can follow either example.

If this were production code the standards would be higher but it's just a little option for stronger linting. I expect our linting to get refactored more soon so don't want to create more commands that have to keep working as expected. This env var has the air of something that could go away.