@mhofman 2022-06-06
> All the `lint:*` packages get run by `yarn lint` so that would duplicate.

Right, I forgot about that.

> On yours you know what to do.

No I don't. Me 3 days ago learned that I also need to set `NODE_OPTIONS` or I fail with OOM. Me in a few days will have forgotten. Calling commands with explicit env is fairly unergonomic if these are not documented anywhere (I guess they're documented in the GitHub actions `yml` files, but that's kinda obscure and not where I'd look in the first place).

I suppose I would have less qualms if somehow the `.eslintrc.js` file would somehow automatically set the required `max-old-space-size` if the env to trigger type checks is set.